added Setup feature on MongoTestServer

+++ src/MongoDB.Testing/Mongo/MongoTestServer.cs
@@ -5,13 +5,30 @@ namespace MongoDB.Testing.Mongo
 {
     public class MongoTestServer : IDisposable
     {
-        private readonly ushort _port;
         private readonly IRandomMongoDatabase _randomDatabase;
         private readonly IProcess _process;
         private readonly IFileSystem _fileSystem;
         private readonly string _dataPath;
+        private readonly MongoTestServerMode _mode;
         private bool _disposed;
 
+        /// <summary>
+        /// Constructs a <see cref="MongoTestServer" /> instance without starting a new process.
+        /// </summary>
+        /// <param name="port">The mongod.exe instance port number.</param>
+        internal MongoTestServer(UInt16 port)
+        {
+            _randomDatabase = new DefaultRandomMongoDatabase(port);
+            _mode = MongoTestServerMode.WithoutOwnedProcess;
+        }
+
+        /// <summary>
+        /// Constructs a <see cref="MongoTestServer" /> instance by starting a new process.
+        /// </summary>
+        /// <param name="port">The mongod.exe instance port number.</param>
+        /// <param name="process"></param>
+        /// <param name="fileSystem"></param>
+        /// <param name="dataPath"></param>
         internal MongoTestServer(UInt16 port, IProcess process, IFileSystem fileSystem, string dataPath)
         {
             if (process == null)
@@ -29,11 +46,11 @@ namespace MongoDB.Testing.Mongo
                 throw new ArgumentNullException("dataPath");
             }
 
-            _port = port;
             _process = process;
             _fileSystem = fileSystem;
             _dataPath = dataPath;
             _randomDatabase = new DefaultRandomMongoDatabase(port);
+            _mode = MongoTestServerMode.WithOwnedProcess;
         }
 
         public MongoDatabase Database
@@ -42,19 +59,26 @@ namespace MongoDB.Testing.Mongo
         }
 
         /// <summary>
-        /// Dispose will terminate the created mongod.exe instance and delete the created temp folder.
+        /// Terminates the created mongod.exe instance and deletes the created temp folder if the instance is 
+        /// created through <see cref="Start(ushort)" /> or <see cref="Start(ushort, IMongoExeLocator)"/>. If the instance 
+        /// is created through <see cref="Setup" />, the initialized <see cref="IRandomMongoDatabase" /> 
+        /// implementation instance will be dropped.
         /// </summary>
         public void Dispose()
         {
             if (_disposed == false)
             {
                 _randomDatabase.Dispose();
-                _process.Kill();
-                _process.WaitForExit();
 
-                if (_fileSystem.DirectoryExists(_dataPath))
+                if (_mode == MongoTestServerMode.WithOwnedProcess)
                 {
-                    _fileSystem.DeleteDirectory(_dataPath, true);
+                    _process.Kill();
+                    _process.WaitForExit();
+
+                    if (_fileSystem.DirectoryExists(_dataPath))
+                    {
+                        _fileSystem.DeleteDirectory(_dataPath, true);
+                    }
                 }
 
                 _disposed = true;
@@ -72,6 +96,17 @@ namespace MongoDB.Testing.Mongo
         public static MongoTestServer Start(UInt16 port, IMongoExeLocator mongoExeLocator)
         {
             return Start(port, mongoExeLocator, new DefaultProcessStarter(), new DefaultFileSystem());
+        }
+
+        /// <summary>
+        /// Sets up a <see cref="MongoTestServer" /> instance. Use this static method if you already have 
+        /// a mongod.exe instance running with a known port number.
+        /// </summary>
+        /// <param name="port"></param>
+        /// <returns></returns>
+        public static MongoTestServer Setup(UInt16 port)
+        {
+            return new MongoTestServer(port);
         }
 
         internal static MongoTestServer Start(UInt16 port, IMongoExeLocator mongoExeLocator, IProcessStarter processStarter, IFileSystem fileSystem)
@@ -97,6 +132,12 @@ namespace MongoDB.Testing.Mongo
             IProcess process = exeFacade.Start(mongodExeOptions);
 
             return new MongoTestServer(port, process, fileSystem, dbPath);
+        }
+
+        private enum MongoTestServerMode : byte
+        {
+            WithOwnedProcess = 1,
+            WithoutOwnedProcess = 2
         }
     }
 }
